Adds pacakges for the annotation processor codegen classes.
Just tidying up, there's quite a lot of these classes around so this makes them easier to navigate.
