+++ google-ads-annotation-processing/src/main/java/com/google/ads/googleads/annotations/impl/GoogleAdsCatalogAnnotationProcessor.java
@@ -17,15 +17,15 @@ package com.google.ads.googleads.annotations.impl;
 import static java.lang.Math.max;
 
 import com.google.ads.googleads.annotations.api.VersionDescriptor;
-import com.google.ads.googleads.annotations.impl.generators.AbstractGoogleAdsClientCodeGenerator;
-import com.google.ads.googleads.annotations.impl.generators.CustomerUserAccessMessageProxyGenerator;
-import com.google.ads.googleads.annotations.impl.generators.FeedMessageProxyGenerator;
-import com.google.ads.googleads.annotations.impl.generators.GeneratedCatalogCodeGenerator;
 import com.google.ads.googleads.annotations.impl.generators.Generator;
-import com.google.ads.googleads.annotations.impl.generators.GoogleAdsAllVersionsCodeGenerator;
-import com.google.ads.googleads.annotations.impl.generators.SearchPagedResponseMessageProxyGenerator;
-import com.google.ads.googleads.annotations.impl.generators.SearchStreamResponseMessageProxyGenerator;
-import com.google.ads.googleads.annotations.impl.generators.VersionCatalogCodeGenerator;
+import com.google.ads.googleads.annotations.impl.generators.catalog.AbstractGoogleAdsClientCodeGenerator;
+import com.google.ads.googleads.annotations.impl.generators.catalog.GeneratedCatalogCodeGenerator;
+import com.google.ads.googleads.annotations.impl.generators.catalog.GoogleAdsAllVersionsCodeGenerator;
+import com.google.ads.googleads.annotations.impl.generators.catalog.VersionCatalogCodeGenerator;
+import com.google.ads.googleads.annotations.impl.generators.messageproxy.CustomerUserAccessMessageProxyGenerator;
+import com.google.ads.googleads.annotations.impl.generators.messageproxy.FeedMessageProxyGenerator;
+import com.google.ads.googleads.annotations.impl.generators.messageproxy.SearchPagedResponseMessageProxyGenerator;
+import com.google.ads.googleads.annotations.impl.generators.messageproxy.SearchStreamResponseMessageProxyGenerator;
 import com.google.auto.service.AutoService;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.SortedSetMultimap;
