+++ google-ads-annotation-processing/src/main/java/com/google/ads/googleads/annotations/impl/generators/messageproxy/AbstractMessageProxyGenerator.java
@@ -12,8 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.google.ads.googleads.annotations.impl.generators;
+package com.google.ads.googleads.annotations.impl.generators.messageproxy;
 
+import com.google.ads.googleads.annotations.impl.generators.Generator;
+import com.google.ads.googleads.annotations.impl.generators.Utils;
 import com.squareup.javapoet.ClassName;
 import com.squareup.javapoet.MethodSpec;
 import com.squareup.javapoet.ParameterizedTypeName;
