+++ google-ads-annotation-processing/src/test/java/com/google/ads/googleads/annotations/impl/generators/messageproxy/SearchStreamResponseMessageProxyGeneratorTest.java
@@ -12,10 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.google.ads.googleads.annotations.impl.generators;
+package com.google.ads.googleads.annotations.impl.generators.messageproxy;
 
-import static com.google.ads.googleads.annotations.impl.generators.AbstractMessageProxyGenerator.MIN_VERSION_FOR_CUSTOMER_USER_ACCESS;
+import static com.google.ads.googleads.annotations.impl.generators.messageproxy.AbstractMessageProxyGenerator.MIN_VERSION_FOR_CUSTOMER_USER_ACCESS;
 
+import com.google.ads.googleads.annotations.impl.generators.AbstractGeneratorTest;
 import com.google.common.collect.ImmutableSet;
 import java.io.IOException;
 import org.junit.Test;
